requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
python-dotenv>=0.19.0
gql[all]>=3.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.5.0
langfuse>=2.0.0
aiohttp>=3.8.0